Separate rendering process to able to execute unittest.
